% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTL.R
\name{plot.cvMTL}
\alias{plot.cvMTL}
\title{Plot the cross-validation curve}
\usage{
\method{plot}{cvMTL}(x, ...)
}
\arguments{
\item{x}{The returned object of function \code{cvMTL}}

\item{...}{Other parameters}
}
\description{
Plot the cross-validation curve
}
\examples{
#create the example data
data<-Create_simulated_data(Regularization="L21", type="Classification")
#perform the cv
cvfit<-cvMTL(data$X, data$Y, type="Classification", Regularization="L21", 
    Lam2=0, opts=list(init=0,  tol=10^-6, maxIter=1500), nfolds=5,
    stratify=TRUE, Lam1_seq=10^seq(1,-4, -1))
#plot the curve
plot(cvfit)
}
