% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTL.R
\name{print.MTL}
\alias{print.MTL}
\title{Print the  meta information of the model}
\usage{
\method{print}{MTL}(x, ...)
}
\arguments{
\item{x}{A trained MTL model}

\item{...}{Other parameters}
}
\description{
Print the  meta information of the model
}
\examples{
#create data
data<-Create_simulated_data(Regularization="L21", type="Regression")
#train a MTL model
model<-MTL(data$X, data$Y, type="Regression", Regularization="L21",
    Lam1=0.1, Lam2=0, opts=list(init=0,  tol=10^-6, maxIter=1500)) 
#print the information of the model
print(model)
}
