% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtl.R
\name{rmtl}
\alias{rmtl}
\title{Restricted mean time lost (RMTL) for competing risks adjusted with inverse probability weighted (IPW)}
\usage{
rmtl(
  times,
  event,
  eoi = 1,
  group = NULL,
  weight = NULL,
  tau = NULL,
  alpha = 0.05,
  entry = NULL,
  ...
)
}
\arguments{
\item{times}{time to event (any-cause) or censoring}

\item{event}{integer denoting the event type, where 0=censored. If there is only 1 event type (values take 0,1), results are equivalent to a non-competing risk setting}

\item{eoi}{the event of interest, which should be an event value (defaults to 1).}

\item{group}{exposure/treatment group. If left blank, function calculates the RMTL in the overall sample.}

\item{weight}{user can specify weights to adjust for confounding, i.e. inverse probability weights, overlap weights, or survey weights. If left blank, function calculates unadjusted results.}

\item{tau}{time horizon used to restrict event times. User can supply a clinically meaningful time horizon, otherwise the function calculates the minimum of the maximum event times in each exposure/treatment group}

\item{alpha}{level of significance (defaults to 0.05).}

\item{entry}{optional argument to allow delayed entry times (i.e., age as the time scale)}
}
\value{
IPW (or unadjusted) RMTL in each group, differences between groups, confidence intervals, p-values, and IPW cumulative incidence curves.
}
\description{
Restricted mean time lost (RMTL) for competing risks adjusted with inverse probability weighted (IPW)
}
