% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtl_mod.R
\name{rmtl_mod}
\alias{rmtl_mod}
\title{Regression model for the restricted mean time lost (RMTL) difference estimated wtih inverse probability of censoring (IPCW)}
\usage{
rmtl_mod(times, event, eoi = 1, tau, cov = NULL, strata = FALSE, group = NULL)
}
\arguments{
\item{times}{time to event (any-cause) or censoring}

\item{event}{integer denoting the event type, where 0=censored. If there is only 1 event type (values take 0,1), results are equivalent to a non-competing risk setting}

\item{eoi}{the event of interest, which should be an event value. Defaults to 1.}

\item{tau}{time horizon used to restrict event times}

\item{cov}{matrix of covariates}

\item{strata}{optional argument to stratify censoring weights in subgroups (as desginated in group)}

\item{group}{optional factor to stratify censoring weights in subgroups. Should only be used when strata=TRUE}
}
\value{
estimated regression model for the difference in cause-specific RMTL
}
\description{
Regression model for the restricted mean time lost (RMTL) difference estimated wtih inverse probability of censoring (IPCW)
}
